DX: Update dependencies and config file to use typescript + typechain
