+++ .gitignore
@@ -6,6 +6,7 @@ packages/hardhat/*.txt
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 **/node_modules
 packages/hardhat/artifacts
+packages/hardhat/typechain
 packages/react-app/src/contracts
 packages/hardhat/cache
